// src/api/model/pizzeriaDetailsDto.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * PizzaApp | v1
 * OpenAPI spec version: 1.0.0
 */
import type { PizzeriaAddressDto } from './pizzeriaAddressDto';

export interface PizzeriaDetailsDto {
  id?: string;
  name: string;
  brandName: string;
  phoneNumber: string;
  /** @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$ */
  deliveryCost?: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$ */
  minOrderAmount?: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$ */
  serviceFee?: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  averagePreparationTimeMinutes?: number | string;
  isOpen?: boolean;
  /** @nullable */
  activeMenuId?: string | null;
  address: PizzeriaAddressDto;
}